Implement tree
Implement a tree to hold the directory structure in
order to accommodate not just for file-based
but also directory-based listing.

Add a DirItem also for this.

Make CatalogueItem a anytree NodeMixin.

Change os.walk to go bottom up.

Remove FilterCheck functionality for now since it's
not clear if it's useful like that.